requirements.txt
# Add your project dependencies here
# e.g., requests==2.25.1

# Core
fastapi==0.109.2
uvicorn==0.27.1
python-multipart==0.0.9
python-dotenv==1.0.1

# Image Processing
opencv-python-headless==4.9.0.80
face-recognition==1.3.0
face-recognition-models==0.3.0
Pillow==10.2.0

# OCR
easyocr==1.7.1
pytesseract==0.3.10

# Phone Validation
phonenumbers==8.13.30

# Utils
numpy==1.26.4
python-jose[cryptography]==3.3.0
python-magic==0.4.27
